+++ packages/doars/src/symbols.js
@@ -1,7 +1,6 @@
 export const ATTRIBUTES = Symbol('ATTRIBUTES')
 
 export const COMPONENT = Symbol('COMPONENT')
-export const COMPONENT_SUFFIX = '-state'
 
 export const FOR = Symbol('FOR')
 
@@ -9,9 +8,9 @@ export const IF = Symbol('IF')
 
 export const INITIALIZED = Symbol('INITIALIZED')
 
-export const SYNC_STATE = Symbol('SYNC_STATE')
-
 export const ON = Symbol('ON')
 
 export const REFERENCES = Symbol('REFERENCES')
 export const REFERENCES_CACHE = Symbol('REFERENCES_CACHE')
+
+export const SYNC_STATE = Symbol('SYNC_STATE')